+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Changelog
 Next
 ----
 
+- Add a ``--wait`` flag to ``dcos-docker create`` to also wait for the cluster.
+
 2018.06.12.0
 ------------